% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex_mpShapeList.R
\docType{data}
\name{Ex_mpShapeList}
\alias{Ex_mpShapeList}
\title{example dataset}
\usage{
data(Ex_mpShapeList)
}
\description{
morphomapShape objects from 5 femora
}
\author{
Antonio Profico
}
\keyword{morphomap}
