% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomFem38023.R
\docType{data}
\name{HomFem38023}
\alias{HomFem38023}
\title{example dataset}
\usage{
data(HomFem38023)
}
\description{
3D mesh of a human femur bone
}
\author{
Antonio Profico, Luca Bondioli, Pasquale Raia, Paul O'Higgins, Damiano Marchi
}
\keyword{morphomap}
