% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphomap.R
\docType{package}
\name{morphomap-package}
\alias{morphomap-package}
\alias{morphomap}
\title{2D and 3D cortical thickness maps and cross sectional geometry}
\description{
Tool to process long bone meshes (shape data, morphometric maps and cross-sectional geometry)
}
\author{
Antonio Profico, Luca Bondioli, Pasquale Raia, Paul O'Higgins, Damiano Marchi
}
