% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphomapArray2matrix.R
\name{morphomapArray2matrix}
\alias{morphomapArray2matrix}
\title{morphomapArray2matrix}
\usage{
morphomapArray2matrix(array)
}
\arguments{
\item{array}{an array}
}
\value{
mat a matrix
}
\description{
Convert an array into a matrix
}
\author{
Antonio Profico, Luca Bondioli, Pasquale Raia, Paul O'Higgins, Damiano Marchi
}
