% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphomapCheck.R
\name{morphomapCheck}
\alias{morphomapCheck}
\title{morphomapCheck}
\usage{
morphomapCheck(mesh, col = "white")
}
\arguments{
\item{mesh}{3D mesh: long bone 3D model}

\item{col}{character: color mesh}
}
\description{
Plot the long bone mesh to check the orientation of the long bone
}
\examples{
\donttest{
library(morphomap)
data(HomFem38023)
morphomapCheck(HomFem38023)
}
}
\author{
Antonio Profico, Luca Bondioli, Pasquale Raia, Paul O'Higgins, Damiano Marchi
}
