% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphomapCircle.R
\name{morphomapCircle}
\alias{morphomapCircle}
\title{morphomapCircle}
\usage{
morphomapCircle(r = 1, n = 1000)
}
\arguments{
\item{r}{numeric: radius of the outline}

\item{n}{numeric: number of points along the outline}
}
\value{
mat matrix with coordinates
}
\description{
Define a circular outline
}
\examples{
extsec<-morphomapCircle(10,100)
intsec<-morphomapCircle(8,100)
plot(extsec,asp=1,type="l")
points(intsec,type="l",col=2)
}
\author{
Antonio Profico, Luca Bondioli, Pasquale Raia, Paul O'Higgins, Damiano Marchi
}
