% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphomapExport.R
\name{morphomapExport}
\alias{morphomapExport}
\title{morphomapExport}
\usage{
morphomapExport(mpShapeObject, id, file)
}
\arguments{
\item{mpShapeObject}{list: list containing morphomapShape objects}

\item{id}{character: label name}

\item{file}{character: name the output file}
}
\description{
Export the output from ToothAlignement
}
\author{
Antonio Profico
}
