% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphomapImport.R
\name{morphomapImport}
\alias{morphomapImport}
\title{morphomapImport}
\usage{
morphomapImport(file)
}
\arguments{
\item{file}{character: name of input file}
}
\value{
3D_out num.pointsx3xnum.sect array in which the external outlines are stored

3D_inn num.pointsx3xnum.sect array in which the internal outlines are stored

2D_out num.pointsx2xnum.sect array in which the external outlines are stored

2D_inn num.pointsx2xnum.sect array in which the interal outlines are stored

ALPM_inn array with the coordinates of ALPM coordinates on the external outline

ALPM_out array with the coordinates of ALPM coordinates on the internal outline

mech_length mechanical length of the long bone

start percentage of the mechanical length from which the first section is defined

end percentage of the mechanical length from which the last section is defined
}
\description{
Import a morphomapShape object exported with morphomapExport
}
\author{
Antonio Profico
}
