% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphomapMatrix2array.R
\name{morphomapMatrix2array}
\alias{morphomapMatrix2array}
\title{morphomapMatrix2array}
\usage{
morphomapMatrix2array(matrix, nsects)
}
\arguments{
\item{matrix}{an array}

\item{nsects}{number of cross sections}
}
\value{
array an array
}
\description{
Convert a matrix into an array
}
\author{
Antonio Profico, Luca Bondioli, Pasquale Raia, Paul O'Higgins, Damiano Marchi
}
