% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphomapReadMorphologika.R
\name{morphomapReadMorphologika}
\alias{morphomapReadMorphologika}
\title{morphomapReadMorphologika}
\usage{
morphomapReadMorphologika(file)
}
\arguments{
\item{file}{path of the file to be read}
}
\value{
out list containing an array, labels, groups and variables
}
\description{
Import an array stored in a morphologika file
}
\author{
Antonio Profico, Luca Bondioli, Pasquale Raia, Paul O'Higgins, Damiano Marchi
}
