% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphomapRectangle.R
\name{morphomapRectangle}
\alias{morphomapRectangle}
\title{morphomapRectangle}
\usage{
morphomapRectangle(l = 1, h = 1, n = 1000)
}
\arguments{
\item{l}{numeric: length of the rectangle}

\item{h}{numeric: height of the rectangle}

\item{n}{numeric: number of points along the outline}
}
\value{
mat matrix with coordinates
}
\description{
Define a rectangular outline
}
\examples{
extsec<-morphomapRectangle(10,6,100)
intsec<-morphomapRectangle(8,4,100)
plot(extsec,asp=1,type="l")
points(intsec,type="l",col=2)
}
\author{
Antonio Profico, Luca Bondioli, Pasquale Raia, Paul O'Higgins, Damiano Marchi
}
