% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphomapSort.R
\name{morphomapSort}
\alias{morphomapSort}
\title{morphomapSort}
\usage{
morphomapSort(mat)
}
\arguments{
\item{mat}{numeric matrix: a kx2 matrix}
}
\value{
mat sorted kx2 matrix
}
\description{
Sort a series of points stored as a 2D matrix
}
\examples{
rand<-sample(100)
extsec<-morphomapCircle(10,100)[rand,]
plot(extsec,type="l",asp=1)
sorted<-morphomapSort(extsec)
plot(sorted,type="l",asp=1)
}
\author{
Antonio Profico, Luca Bondioli, Pasquale Raia, Paul O'Higgins, Damiano Marchi
}
