% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphomapTri2sects.R
\name{morphomapTri2sects}
\alias{morphomapTri2sects}
\title{morphomapTri2sects}
\usage{
morphomapTri2sects(cp, mp)
}
\arguments{
\item{cp}{matrix: coordinates of the external outline of the section}

\item{mp}{matrix: coordinates of the internal outline of the section}
}
\value{
matrix coordinates of the triangulated mesh

tri triangulations of the triangulated mesh
}
\description{
Triangulate the external and internal outlines of a 3D cross section
}
\author{
Antonio Profico, Luca Bondioli, Pasquale Raia, Paul O'Higgins, Damiano Marchi
}
