% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphomapTriangulate.R
\name{morphomapTriangulate}
\alias{morphomapTriangulate}
\title{morphomapTriangulate}
\usage{
morphomapTriangulate(set, n, close = FALSE)
}
\arguments{
\item{set}{matrix: coordinates of the cross sections to be triangulated}

\item{n}{numeric: number of cross sections}

\item{close}{logical: if TRUE the two surfaces are closed}
}
\value{
mesh a mesh of the triangulated semilandark configuration
}
\description{
Build a mesh starting from the coordinates of the diaphysis
}
\author{
Antonio Profico, Luca Bondioli, Pasquale Raia, Paul O'Higgins, Damiano Marchi
}
