% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphomapWriteMorphologika.R
\name{morphomapWriteMorphologika}
\alias{morphomapWriteMorphologika}
\title{morphomapWriteMorphologika}
\usage{
morphomapWriteMorphologika(array, groups = NULL, variables = NULL, file)
}
\arguments{
\item{array}{an array}

\item{groups}{a vector containing a classifier}

\item{variables}{list containing further classifiers}

\item{file}{path of the file to be saved}
}
\description{
Export an array in the morphologika format file
}
\author{
Antonio Profico, Luca Bondioli, Pasquale Raia, Paul O'Higgins, Damiano Marchi
}
