% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanFem27713.R
\docType{data}
\name{PanFem27713}
\alias{PanFem27713}
\title{example dataset}
\usage{
data(PanFem27713)
}
\description{
3D mesh of a chimpanzee femur bone
}
\author{
Antonio Profico, Luca Bondioli, Pasquale Raia, Paul O'Higgins, Damiano Marchi
}
\keyword{morphomap}
